pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cicd-diagnosis"
version = "1.0.0"
description = "Automated CI/CD failure diagnosis using LLMs"
authors = [{name = "Fahid Khan", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "chromadb>=0.4.18",
    "sentence-transformers>=2.2.2",
    "scikit-learn>=1.3.2",
    "scipy>=1.11.4",
    "matplotlib>=3.8.2",
    "seaborn>=0.13.0",
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.1",
    "pandas>=2.1.3",
    "numpy>=1.26.2",
    "pyyaml>=6.0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 120
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false